+++ tests/test_utils.py
@@ -3,6 +3,7 @@ from datetime import datetime
 from importlib import resources
 
 import pandas as pd
+from profiler import __version__ as pv
 
 from claspar import __version__, utils
 
@@ -101,13 +102,18 @@ def test_create_bacterial_analysis_fields():
     today = datetime.today().strftime("%Y-%m-%d")
     expected_analysis_table_dict = {
         "identifiers": [],
-        "name": "bacteria-classifier-parser",
-        "description": "This is an analysis to parse and filter the bacteria classifications from sylph",
+        "name": "claspar-sylph-bacteria",
+        "description": (
+            "This is an analysis to parse and filter the bacteria classifications from sylph and look up "
+            "the clinical profiles for classified taxa."
+        ),
         "analysis_date": today,
         "pipeline_name": "ClasPar",
         "pipeline_version": __version__,
         "pipeline_url": "https://github.com/ukhsa-collaboration/gpha-mscape-orangebox-claspar",
-        "methods": '{"stuff": 1}',
+        "methods": (
+            '{"stuff": 1, "profiler_version": "%s", "profile_tables_version": "profile_tables_testing.xlsx"}' % pv
+        ),
         "result": "Found some stuff here.",
         "result_metrics": '{"0": {"thing": 10, "type": "little"}, "1": {"thing": 10, "type": "big"}}',
         "server_records": ["ID_123456"],
@@ -117,6 +123,7 @@ def test_create_bacterial_analysis_fields():
         classifier="sylph",
         record_id="ID_123456",
         thresholds={"stuff": 1},
+        profile_table_name="profile_tables_testing.xlsx",
         headline_result="Found some stuff here.",
         results={0: {"thing": 10, "type": "little"}, 1: {"thing": 10, "type": "big"}},
         server="server",